Relocate generic ESP-Xtensa boot infra to chip-neutral module
Move rom_thunks (RomThunkBank, APPCPU bring-up flags, RomThunkFn) and
system_stub (SystemStub/EfuseStub/RtcCntlStub) out of peripherals/esp32s3/
into a new chip-neutral peripherals/esp_xtensa_common/ module. These are
generic ESP-Xtensa bring-up plumbing and peripheral-as-RAM stubs, not S3
silicon, so the classic ESP32 + config layer no longer import from another
chip.

Update all importers (lib, bus, cpu/xtensa_lx7, system/xtensa,
esp32s3 {core1_control, system, wifi_thunks}, wasm, cli, tests) and the
arch_chip_isolation guard's ALLOWED file-suffix path.
