+++ crates/core/src/bus/mod.rs
@@ -158,7 +158,7 @@ pub struct SystemBus {
     /// dispatcher (xtensa_lx7.rs) calls `bus.get_rom_thunk(pc)` which
     /// checks both this table and any `RomThunkBank` peripherals.
     pub flash_thunks:
-        std::collections::HashMap<u32, crate::peripherals::esp32s3::rom_thunks::RomThunkFn>,
+        std::collections::HashMap<u32, crate::peripherals::esp_xtensa_common::rom_thunks::RomThunkFn>,
     peripheral_ranges: Vec<PeripheralRange>,
     peripheral_hint: Cell<Option<usize>>,
     /// Cached index of the classic-ESP32 DPORT peripheral, if one is
@@ -965,7 +965,7 @@ impl SystemBus {
     pub fn get_rom_thunk(
         &self,
         pc: u32,
-    ) -> Option<crate::peripherals::esp32s3::rom_thunks::RomThunkFn> {
+    ) -> Option<crate::peripherals::esp_xtensa_common::rom_thunks::RomThunkFn> {
         // First check the Bus-level flash thunk table (for thunks installed
         // outside any RomThunkBank's range — typically firmware functions
         // resident in flash that we want to intercept).
@@ -978,7 +978,7 @@ impl SystemBus {
             if pc >= base && pc < end {
                 if let Some(any) = p.dev.as_any() {
                     if let Some(bank) =
-                        any.downcast_ref::<crate::peripherals::esp32s3::rom_thunks::RomThunkBank>()
+                        any.downcast_ref::<crate::peripherals::esp_xtensa_common::rom_thunks::RomThunkBank>()
                     {
                         return bank.get(pc);
                     }
@@ -996,9 +996,9 @@ impl SystemBus {
     pub fn install_flash_thunk(
         &mut self,
         pc: u32,
-        thunk: crate::peripherals::esp32s3::rom_thunks::RomThunkFn,
+        thunk: crate::peripherals::esp_xtensa_common::rom_thunks::RomThunkFn,
     ) -> SimResult<()> {
-        let bytes = crate::peripherals::esp32s3::rom_thunks::ROM_THUNK_BREAK_BYTES;
+        let bytes = crate::peripherals::esp_xtensa_common::rom_thunks::ROM_THUNK_BREAK_BYTES;
         for (i, b) in bytes.iter().enumerate() {
             self.write_u8(pc as u64 + i as u64, *b)?;
         }
@@ -2943,7 +2943,7 @@ impl crate::Bus for SystemBus {
     fn get_rom_thunk(
         &self,
         pc: u32,
-    ) -> Option<crate::peripherals::esp32s3::rom_thunks::RomThunkFn> {
+    ) -> Option<crate::peripherals::esp_xtensa_common::rom_thunks::RomThunkFn> {
         SystemBus::get_rom_thunk(self, pc)
     }
