+++ crates/core/src/peripherals/esp32s3/mod.rs
@@ -24,7 +24,6 @@ pub mod mcpwm;
 pub mod pcnt;
 pub mod rmt;
 pub mod rng;
-pub mod rom_thunks;
 pub mod rsa;
 pub mod sar_adc;
 pub mod sdmmc;
@@ -32,7 +31,6 @@ pub mod sens;
 pub mod sha;
 pub mod spi_mem_flash;
 pub mod system;
-pub mod system_stub;
 pub mod systimer;
 pub mod timer_group;
 pub mod tmp102;
