+++ crates/core/src/peripherals/esp_xtensa_common/rom_thunks.rs
